refactor: update TypeScript configuration and enhance environment variable definitions
