+++ src/components/navBar.vue
@@ -2,7 +2,7 @@
 nav#navBar(ref="navBar")
     .wrap
         .left
-            router-link.logo(to="/my-services" v-if="route.name != 'home' && loginState && route.path !== '/my-services'" style="color:white")
+            router-link.logo(to="/my-services" v-if="route.name != 'home' && loggedAccount && route.path !== '/my-services'" style="color:white")
                 //- .material-symbols-outlined.notranslate.nohover.back(style="font-size:1.5em") arrow_back_ios
                 svg(width="1.5em" height="1.5em" style="fill:white")
                     use(xlink:href="@/assets/img/material-icon.svg#icon-arrow-back-ios")
@@ -12,7 +12,7 @@ nav#navBar(ref="navBar")
                 img.symbol.desktop(src="@/assets/img/logo/logo-white.svg" style="image-orientation: none;height:38px")
         .right
             ul.menu-wrap
-                template(v-if="loginState")
+                template(v-if="loggedAccount")
                     li.go-github
                         a(href="https://github.com/broadwayinc/skapi-js" target="_blank")
                             img(src="@/assets/img/icon/icon_github.svg")
@@ -76,8 +76,9 @@ nav#navBar(ref="navBar")
 <script setup lang="ts">
 import { useRoute, useRouter } from "vue-router";
 import { onMounted, onBeforeUnmount, ref } from "vue";
-import { skapi } from "@/main";
-import { loginState, user, updateUser, customer } from "@/code/user";
+import { skapi, loggedAccount } from "@/main";
+// import { loginState, user, updateUser, customer } from "@/code/user";
+import { user, updateUser, customer } from "@/code/user";
 import { showDropDown } from "@/assets/js/event.js";
 import { setAutoHide, removeListener } from "./navBar-autohide.ts";
