+++ src/views/service/main.vue
@@ -142,7 +142,7 @@ div(v-else style='text-align: center;margin-top: 100px;')
 <script setup lang="ts">
 import { useRoute, useRouter } from "vue-router";
 import { ref, watch } from "vue";
-import { loginState } from "@/code/user";
+// import { loginState } from "@/code/user";
 import { serviceList } from "@/views/service-list";
 import { currentService, setService, serviceMainLoaded } from "@/views/service/main";
 import Modal from "@/components/modal.vue";
@@ -181,15 +181,15 @@ let prevRouter = ref("");
 let nextRouter = ref("");
 let plan = ref("Trial");
 
-watch(
-    loginState,
-    (nv) => {
-        if (!nv) {
-            router.push("/login");
-        }
-    },
-    { immediate: true }
-);
+// watch(
+//     loginState,
+//     (nv) => {
+//         if (!nv) {
+//             router.push("/login");
+//         }
+//     },
+//     { immediate: true }
+// );
 
 watch(serviceList, nv => {
     if (nv[serviceId] && currentService?.id !== serviceId) {
